Refactor: Move Selenium test components to a base class
